Migrate ClassDB to LuaLazyCache
