--- runtime/data/class.lua
@@ -1,359 +1,0 @@
-class.warrior = {
-  ordering = 10010,
-  item_type = 1,
-  is_extra = false,
-  equipment_type = 1,
-  skills = {
-    _10 = 10,
-    _11 = 10,
-    _12 = 2,
-    _13 = 0,
-    _14 = 0,
-    _15 = 3,
-    _16 = 0,
-    _17 = 0,
-    _18 = 0,
-    _100 = 6,
-    _101 = 4,
-    _103 = 6,
-    _102 = 6,
-    _104 = 4,
-    _107 = 5,
-    _167 = 6,
-    _152 = 4,
-    _173 = 5,
-    _154 = 5,
-    _170 = 4,
-    _169 = 4,
-    _168 = 5,
-  },
-}
-class.thief = {
-  ordering = 10020,
-  item_type = 7,
-  is_extra = false,
-  equipment_type = 7,
-  skills = {
-    _10 = 4,
-    _11 = 4,
-    _12 = 8,
-    _13 = 5,
-    _14 = 4,
-    _15 = 0,
-    _16 = 0,
-    _17 = 0,
-    _18 = 0,
-    _101 = 4,
-    _100 = 3,
-    _108 = 3,
-    _173 = 4,
-    _166 = 4,
-    _152 = 3,
-    _156 = 4,
-    _300 = 4,
-    _174 = 3,
-    _170 = 4,
-  },
-}
-class.wizard = {
-  ordering = 10030,
-  item_type = 2,
-  is_extra = false,
-  equipment_type = 2,
-  skills = {
-    _10 = 0,
-    _11 = 0,
-    _12 = 0,
-    _13 = 4,
-    _14 = 3,
-    _15 = 8,
-    _16 = 10,
-    _17 = 0,
-    _18 = 0,
-    _105 = 3,
-    _101 = 2,
-    _150 = 6,
-    _165 = 3,
-    _164 = 6,
-    _174 = 5,
-    _171 = 4,
-    _178 = 4,
-    _172 = 5,
-  },
-}
-class.farmer = {
-  ordering = 10040,
-  item_type = 1,
-  is_extra = false,
-  equipment_type = 1,
-  skills = {
-    _10 = 5,
-    _11 = 5,
-    _12 = 2,
-    _13 = 0,
-    _14 = 8,
-    _15 = 5,
-    _16 = 0,
-    _17 = 0,
-    _18 = 0,
-    _104 = 4,
-    _107 = 3,
-    _173 = 3,
-    _156 = 5,
-    _184 = 6,
-    _161 = 7,
-    _154 = 5,
-    _180 = 5,
-    _177 = 5,
-  },
-}
-class.predator = {
-  ordering = 20010,
-  item_type = 0,
-  is_extra = true,
-  equipment_type = 0,
-  skills = {
-    _10 = 8,
-    _11 = 11,
-    _12 = 8,
-    _13 = 0,
-    _14 = 0,
-    _15 = 0,
-    _16 = 0,
-    _17 = 0,
-    _18 = 10,
-    _152 = 4,
-  },
-}
-class.archer = {
-  ordering = 10050,
-  item_type = 3,
-  is_extra = false,
-  equipment_type = 3,
-  skills = {
-    _10 = 6,
-    _11 = 4,
-    _12 = 8,
-    _13 = 5,
-    _14 = 2,
-    _15 = 0,
-    _16 = 0,
-    _17 = 0,
-    _18 = 0,
-    _108 = 5,
-    _109 = 5,
-    _101 = 4,
-    _102 = 3,
-    _173 = 5,
-    _174 = 3,
-    _170 = 3,
-    _177 = 4,
-    _301 = 4,
-    _189 = 3,
-  },
-}
-class.warmage = {
-  ordering = 10060,
-  item_type = 5,
-  is_extra = false,
-  equipment_type = 5,
-  skills = {
-    _10 = 6,
-    _11 = 6,
-    _12 = 2,
-    _13 = 0,
-    _14 = 0,
-    _15 = 4,
-    _16 = 7,
-    _17 = 0,
-    _18 = 0,
-    _100 = 4,
-    _101 = 3,
-    _173 = 3,
-    _150 = 4,
-    _164 = 3,
-    _174 = 5,
-    _170 = 4,
-    _169 = 4,
-    _172 = 4,
-  },
-}
-class.tourist = {
-  ordering = 10070,
-  item_type = 0,
-  is_extra = false,
-  equipment_type = 0,
-  skills = {
-    _10 = 0,
-    _11 = 0,
-    _12 = 0,
-    _13 = 0,
-    _14 = 0,
-    _15 = 0,
-    _16 = 0,
-    _17 = 0,
-    _18 = 0,
-    _185 = 5,
-    _182 = 3,
-  },
-}
-class.pianist = {
-  ordering = 10080,
-  item_type = 3,
-  is_extra = false,
-  equipment_type = 3,
-  skills = {
-    _10 = 6,
-    _11 = 0,
-    _12 = 4,
-    _13 = 5,
-    _14 = 6,
-    _15 = 0,
-    _16 = 4,
-    _17 = 8,
-    _18 = 0,
-    _183 = 6,
-    _153 = 19,
-    _150 = 4,
-    _165 = 6,
-    _174 = 6,
-    _179 = 5,
-    _171 = 4,
-    _301 = 3,
-  },
-}
-class.gunner = {
-  ordering = 20020,
-  item_type = 4,
-  is_extra = true,
-  equipment_type = 4,
-  skills = {
-    _10 = 0,
-    _11 = 2,
-    _12 = 5,
-    _13 = 8,
-    _14 = 5,
-    _15 = 4,
-    _16 = 3,
-    _17 = 0,
-    _18 = 0,
-    _110 = 5,
-    _173 = 4,
-    _150 = 3,
-    _154 = 4,
-    _189 = 3,
-  },
-}
-class.priest = {
-  ordering = 10090,
-  item_type = 6,
-  is_extra = false,
-  equipment_type = 6,
-  skills = {
-    _10 = 2,
-    _11 = 2,
-    _12 = 0,
-    _13 = 2,
-    _14 = 2,
-    _15 = 10,
-    _16 = 7,
-    _17 = 2,
-    _18 = 0,
-    _103 = 3,
-    _168 = 3,
-    _150 = 5,
-    _154 = 5,
-    _174 = 5,
-    _170 = 3,
-    _169 = 4,
-    _181 = 5,
-    _172 = 5,
-  },
-}
-class.claymore = {
-  ordering = 10100,
-  item_type = 8,
-  is_extra = false,
-  equipment_type = 8,
-  skills = {
-    _10 = 9,
-    _11 = 3,
-    _12 = 7,
-    _13 = 6,
-    _14 = 0,
-    _15 = 0,
-    _16 = 4,
-    _17 = 0,
-    _18 = 0,
-    _100 = 6,
-    _167 = 7,
-    _152 = 5,
-    _173 = 7,
-    _187 = 4,
-    _154 = 6,
-    _171 = 5,
-    _150 = 4,
-  },
-}
-
-
--- For debug
-class.debugger = {
-  ordering = 0,
-  item_type = 1,
-  is_extra = true,
-  equipment_type = 1,
-  skills = {
-    _10 = 50,
-    _11 = 50,
-    _12 = 50,
-    _13 = 50,
-    _14 = 50,
-    _15 = 50,
-    _16 = 50,
-    _17 = 50,
-    _18 = 50,
-    _100 = 50,
-    _101 = 50,
-    _102 = 50,
-    _103 = 50,
-    _104 = 50,
-    _105 = 50,
-    _107 = 50,
-    _108 = 50,
-    _109 = 50,
-    _110 = 50,
-    _150 = 50,
-    _152 = 50,
-    _153 = 509,
-    _154 = 50,
-    _156 = 50,
-    _161 = 50,
-    _164 = 50,
-    _165 = 50,
-    _166 = 50,
-    _167 = 50,
-    _168 = 50,
-    _169 = 50,
-    _170 = 50,
-    _171 = 50,
-    _172 = 50,
-    _173 = 50,
-    _174 = 50,
-    _177 = 50,
-    _178 = 50,
-    _179 = 50,
-    _180 = 50,
-    _181 = 50,
-    _182 = 50,
-    _183 = 50,
-    _184 = 50,
-    _185 = 50,
-    _187 = 50,
-    _189 = 50,
-    _300 = 50,
-    _301 = 50,
-  },
-}
-
--- vim: et sw=2 sts=2
