+++ runtime/mods/core/data.lua
@@ -14,3 +14,4 @@ require("data/ability")
 require("data/god")
 require("data/item_material")
 require("data/race")
+require("data/class")
