+++ runtime/mods/core/data/class.lua
@@ -1,0 +1,373 @@
+data:define_type("class")
+data:add_multi(
+   "core.class",
+   {
+      {
+         name = "warrior",
+         ordering = 10010,
+         item_type = 1,
+         is_extra = false,
+         equipment_type = 1,
+         skills = {
+            [10] = 10,
+            [11] = 10,
+            [12] = 2,
+            [13] = 0,
+            [14] = 0,
+            [15] = 3,
+            [16] = 0,
+            [17] = 0,
+            [18] = 0,
+            [100] = 6,
+            [101] = 4,
+            [103] = 6,
+            [102] = 6,
+            [104] = 4,
+            [107] = 5,
+            [167] = 6,
+            [152] = 4,
+            [173] = 5,
+            [154] = 5,
+            [170] = 4,
+            [169] = 4,
+            [168] = 5,
+         },
+      },
+      {
+         name = "thief",
+         ordering = 10020,
+         item_type = 7,
+         is_extra = false,
+         equipment_type = 7,
+         skills = {
+            [10] = 4,
+            [11] = 4,
+            [12] = 8,
+            [13] = 5,
+            [14] = 4,
+            [15] = 0,
+            [16] = 0,
+            [17] = 0,
+            [18] = 0,
+            [101] = 4,
+            [100] = 3,
+            [108] = 3,
+            [173] = 4,
+            [166] = 4,
+            [152] = 3,
+            [156] = 4,
+            [300] = 4,
+            [174] = 3,
+            [170] = 4,
+         },
+      },
+      {
+         name = "wizard",
+         ordering = 10030,
+         item_type = 2,
+         is_extra = false,
+         equipment_type = 2,
+         skills = {
+            [10] = 0,
+            [11] = 0,
+            [12] = 0,
+            [13] = 4,
+            [14] = 3,
+            [15] = 8,
+            [16] = 10,
+            [17] = 0,
+            [18] = 0,
+            [105] = 3,
+            [101] = 2,
+            [150] = 6,
+            [165] = 3,
+            [164] = 6,
+            [174] = 5,
+            [171] = 4,
+            [178] = 4,
+            [172] = 5,
+         },
+      },
+      {
+         name = "farmer",
+         ordering = 10040,
+         item_type = 1,
+         is_extra = false,
+         equipment_type = 1,
+         skills = {
+            [10] = 5,
+            [11] = 5,
+            [12] = 2,
+            [13] = 0,
+            [14] = 8,
+            [15] = 5,
+            [16] = 0,
+            [17] = 0,
+            [18] = 0,
+            [104] = 4,
+            [107] = 3,
+            [173] = 3,
+            [156] = 5,
+            [184] = 6,
+            [161] = 7,
+            [154] = 5,
+            [180] = 5,
+            [177] = 5,
+         },
+      },
+      {
+         name = "predator",
+         ordering = 20010,
+         item_type = 0,
+         is_extra = true,
+         equipment_type = 0,
+         skills = {
+            [10] = 8,
+            [11] = 11,
+            [12] = 8,
+            [13] = 0,
+            [14] = 0,
+            [15] = 0,
+            [16] = 0,
+            [17] = 0,
+            [18] = 10,
+            [152] = 4,
+         },
+      },
+      {
+         name = "archer",
+         ordering = 10050,
+         item_type = 3,
+         is_extra = false,
+         equipment_type = 3,
+         skills = {
+            [10] = 6,
+            [11] = 4,
+            [12] = 8,
+            [13] = 5,
+            [14] = 2,
+            [15] = 0,
+            [16] = 0,
+            [17] = 0,
+            [18] = 0,
+            [108] = 5,
+            [109] = 5,
+            [101] = 4,
+            [102] = 3,
+            [173] = 5,
+            [174] = 3,
+            [170] = 3,
+            [177] = 4,
+            [301] = 4,
+            [189] = 3,
+         },
+      },
+      {
+         name = "warmage",
+         ordering = 10060,
+         item_type = 5,
+         is_extra = false,
+         equipment_type = 5,
+         skills = {
+            [10] = 6,
+            [11] = 6,
+            [12] = 2,
+            [13] = 0,
+            [14] = 0,
+            [15] = 4,
+            [16] = 7,
+            [17] = 0,
+            [18] = 0,
+            [100] = 4,
+            [101] = 3,
+            [173] = 3,
+            [150] = 4,
+            [164] = 3,
+            [174] = 5,
+            [170] = 4,
+            [169] = 4,
+            [172] = 4,
+         },
+      },
+      {
+         name = "tourist",
+         ordering = 10070,
+         item_type = 0,
+         is_extra = false,
+         equipment_type = 0,
+         skills = {
+            [10] = 0,
+            [11] = 0,
+            [12] = 0,
+            [13] = 0,
+            [14] = 0,
+            [15] = 0,
+            [16] = 0,
+            [17] = 0,
+            [18] = 0,
+            [185] = 5,
+            [182] = 3,
+         },
+      },
+      {
+         name = "pianist",
+         ordering = 10080,
+         item_type = 3,
+         is_extra = false,
+         equipment_type = 3,
+         skills = {
+            [10] = 6,
+            [11] = 0,
+            [12] = 4,
+            [13] = 5,
+            [14] = 6,
+            [15] = 0,
+            [16] = 4,
+            [17] = 8,
+            [18] = 0,
+            [183] = 6,
+            [153] = 19,
+            [150] = 4,
+            [165] = 6,
+            [174] = 6,
+            [179] = 5,
+            [171] = 4,
+            [301] = 3,
+         },
+      },
+      {
+         name = "gunner",
+         ordering = 20020,
+         item_type = 4,
+         is_extra = true,
+         equipment_type = 4,
+         skills = {
+            [10] = 0,
+            [11] = 2,
+            [12] = 5,
+            [13] = 8,
+            [14] = 5,
+            [15] = 4,
+            [16] = 3,
+            [17] = 0,
+            [18] = 0,
+            [110] = 5,
+            [173] = 4,
+            [150] = 3,
+            [154] = 4,
+            [189] = 3,
+         },
+      },
+      {
+         name = "priest",
+         ordering = 10090,
+         item_type = 6,
+         is_extra = false,
+         equipment_type = 6,
+         skills = {
+            [10] = 2,
+            [11] = 2,
+            [12] = 0,
+            [13] = 2,
+            [14] = 2,
+            [15] = 10,
+            [16] = 7,
+            [17] = 2,
+            [18] = 0,
+            [103] = 3,
+            [168] = 3,
+            [150] = 5,
+            [154] = 5,
+            [174] = 5,
+            [170] = 3,
+            [169] = 4,
+            [181] = 5,
+            [172] = 5,
+         },
+      },
+      {
+         name = "claymore",
+         ordering = 10100,
+         item_type = 8,
+         is_extra = false,
+         equipment_type = 8,
+         skills = {
+            [10] = 9,
+            [11] = 3,
+            [12] = 7,
+            [13] = 6,
+            [14] = 0,
+            [15] = 0,
+            [16] = 4,
+            [17] = 0,
+            [18] = 0,
+            [100] = 6,
+            [167] = 7,
+            [152] = 5,
+            [173] = 7,
+            [187] = 4,
+            [154] = 6,
+            [171] = 5,
+            [150] = 4,
+         },
+      },
+      -- For debug
+      {
+         name = "debugger",
+         ordering = 0,
+         item_type = 1,
+         is_extra = true,
+         equipment_type = 1,
+         skills = {
+            [10] = 50,
+            [11] = 50,
+            [12] = 50,
+            [13] = 50,
+            [14] = 50,
+            [15] = 50,
+            [16] = 50,
+            [17] = 50,
+            [18] = 50,
+            [100] = 50,
+            [101] = 50,
+            [102] = 50,
+            [103] = 50,
+            [104] = 50,
+            [105] = 50,
+            [107] = 50,
+            [108] = 50,
+            [109] = 50,
+            [110] = 50,
+            [150] = 50,
+            [152] = 50,
+            [153] = 509,
+            [154] = 50,
+            [156] = 50,
+            [161] = 50,
+            [164] = 50,
+            [165] = 50,
+            [166] = 50,
+            [167] = 50,
+            [168] = 50,
+            [169] = 50,
+            [170] = 50,
+            [171] = 50,
+            [172] = 50,
+            [173] = 50,
+            [174] = 50,
+            [177] = 50,
+            [178] = 50,
+            [179] = 50,
+            [180] = 50,
+            [181] = 50,
+            [182] = 50,
+            [183] = 50,
+            [184] = 50,
+            [185] = 50,
+            [187] = 50,
+            [189] = 50,
+            [300] = 50,
+            [301] = 50,
+         },
+      },
+})
