+++ src/data/init.cpp
@@ -39,6 +39,9 @@ void _initialize_iterable_dbs(lua::DataTable& data)
 
     the_race_db.initialize(data);
     the_race_db.load_all();
+
+    the_class_db.initialize(data);
+    the_class_db.load_all();
 }
