+++ src/data/types.cpp
@@ -1,6 +1,7 @@
 #include "types/type_ability.cpp"
 #include "types/type_chara_chip.cpp"
 #include "types/type_character.cpp"
+#include "types/type_class.cpp"
 #include "types/type_fish.cpp"
 #include "types/type_god.cpp"
 #include "types/type_item.cpp"
