+++ src/data/types.hpp
@@ -2,6 +2,7 @@
 #include "types/type_ability.hpp"
 #include "types/type_chara_chip.hpp"
 #include "types/type_character.hpp"
+#include "types/type_class.hpp"
 #include "types/type_fish.hpp"
 #include "types/type_god.hpp"
 #include "types/type_item.hpp"
